Add BuildTime and BuildVersion to http-connector healthCheck
